refactor(sample): migrate sample app to 'best practice' directory layout
The current example is a bit too simplistic in my experience to fully grasp
the power of ui-router. It provides a monolithic example of state definition
that shows the varied approaches that can be used, but not in a very structured
format. This change applies the emerging best practice directory layout of an
angular application to exemplify the power of ui-router