improve(uiSref+params): experimental improvements
Let ui-sref understand params without object notation, assuming an object which key is the same as the varname; e.g.
  ui-sref="sign-in(credentials, verify)"
is the same as
  ui-sref="sign-in({credentials:credentials, verify:verify})"

State params are not normalized to String, but they remain as object or whatever type they are.

State params are not limited to the ones defined in the url or params array. Any param passed to transitionTo method is passed to the state.

State params in url and params array are not exclusive, both can be used.

Binds to the submit event of the form if the state doesn't have an url.