Update library to export main module through CommonJS
The convention for exporting data with angular modules is to export the main module. This this even more important as ES6 is being adopted and the convention is to import modules which use a default export to expose the main module, then referring to the module's name using the 'name' property from the angular module API.  Since only the name of this module was exported, this would prevent the use of the module in the usual way and force you to hardcode the dependency name. This fix allows you to do:

```
import ModuleRouter from 'angular-ui-router';

let MyModule = angular.module('MyModule', [ModuleRouter.name]);

...

exports default MyModule;

```